Add files via upload
A drunk at the helm of a scout craft 'THE CALAMITY', crashes onto a desolate planetoid. The 'CALAMITY' lies immobile, a victim of a magnetic force designed to remove space debris. The human, trapped, finds himself among sentient machines, abandoned and swept up by unknown races. These machines, (THE SYNTHOID) seek to escape the planet using the scout craft, but they're imprisoned by the very device (THE SENTINEL:NULL MACHINE)that sucked the craft in. The ‘SYNTHOID’ are powerless against the 'SENTINEL'. The human, however, as an 'ORGANIC BEING', can get close enough to disable it. A fragile alliance forms between the human and the cunning machines. The human is wary, suspecting he will be discarded once the machines gain control of his craft. But the machines are stymied by the craft's fail safe system. Activated  the 'SENTINELS SUICIDE DRONES' are determined to stop them.